requirements.dev.txt
django-debug-toolbar>=3.2.4
python-dotenv==0.20.0